Rewrite how async comms work between the components.
Instead of forcing users to run a select loop and read from all the
channels to avoid locking everything, provide a simple interface that
they can implement and have the client run the loop while passing the
events on to whoever registered as a listener.

We still probably want another API that offers lower level control
of this, but that's later.