Add unit tests for GradualRolloutUserId.
These unit tests are based on the Node.js tests for the same type. The pattern
in Go is to have one test per function and then several sub-tests to test
variations of the input. The names of the tests have no significance, they are
just a shorthand notation that should make it easier to identify them.