Introduce a WaitForReady function to block until the client is ready
Depending on the speed of the repository storage implementation, it
can take a bit of time before the repository is fully ready. Calling
IsEnabled before this happens can result in incorrect results. We
needed a way to delay calling IsEnabled until the client is ready so
add a function that will block until that happens or simply returns
immediately if the client is already loaded.