Fix conversion of uint32 to float64
For doing the rollout calculations we convert everything to float64, but
we check the type before doing the conversion so it is somewhat fragile.
When we switched to using Murmur3, we switched the return type of that
function to be a uint32 and it fell through the cracks of this
conversion function.