Add support for CIDR ranges in lists of IPs
This allows the server to use CIDR ranges when specifying the list of
valid IP addresses for a feature. This also improved the testing for
normal IP addresses because now we are parsing them into a proper IP
address struct and using the builtin `Equal` function. This has an
added benefit over pure string matching because if the remote address is
IPv6, we can still potentially match against an IPv4 address or range
and vice versa.

Fixes #13