Make v3 have it's own import path
As recommended by the new Go module system, major versions that are not
compatible with previous versions should get their own import path.

https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher

While the API of the module is still compatible, we can't just make
the master branch become version 3 because everyone using the old import
path would suddenly get a client that might not be compatible with their
server if they are running an older version. It is therefore preferable
to use a separate branch for the v3 client and tag future releases from
this branch.