Remove an unstable test
This test has been failing sporadically on Travis and I haven't been
able to track down the cause because I can't reproduce it locally.
I'm going to remove this for now such that we can merge the PR
containing the client specification test which is more important
in my opinion.