Fix 'WithDisableMetrics' to actually stop sending metrics
This was previously broken because we were never passing down the
boolean value from the client to the metrics struct. Doing so implies
that the timer is never created so we can't read from the channel in the
sync function. If metrics are disabled, there is really no reason to
even run the metrics routine, so just return instead.

Fixes #15