[API CHANGE] Move feature options to a sub-package
With the upcoming variants functionality, some of these options such as
WithContext() will be duplicated and since Go does not support function
overloading, we need a way to "namespace" these functions. One option is
to use a different name for the function (eg: WithVariantContext), but
it feels cleaner to separate this out into a sub-package.