[API CHANGE] Replace DebugListener with DefaultListener
It was a constant source of error that people forgot to supply a
listener when creating the client so instead, let's rebrand the
DebugListener as DefaultListener and make it the default for the client.
It logs all message using the standard logger. The API require no
changes for users with custom listeners, however people wishing to read
from the channels themselves must now use WithListener(nil).

Also start updating the docs with some information about these API
changes.