spec_test: Make sure test fails in case of errors from Unleash client
Before this change a MockedListener was passed to WithListener instead
of a pointer to a MockedListener. MockedListener doesn't implement any of
the *Listener interfaces as it has pointer receivers on all its methods.
The problem was hidden because WithListener takes an interface{} as
argument and then uses type assertions to check if the argument implements
the *Listener interfaces.