Allow generating random numbers safely from multiple Go routines
Reading from math/rand from multiple Go routines concurrently is not
safe so we need to protect the reads with a mutex. Also move all
convenience functions to be on the new struct. The LockedSource
implementation from the 'exp' package was considered, but it has no
exported fields so could not be initialized from the outside.