TestMetrics_SendMetricsFail: Fix non-deterministic test failure
The test failure can be made deterministic by adding this sleep:

    diff --git a/metrics_test.go b/metrics_test.go
    index 68345c3..47ea377 100644
    --- a/metrics_test.go
    +++ b/metrics_test.go
    @@ -206,6 +206,7 @@ func TestMetrics_SendMetricsFail(t *testing.T) {
            // /client/metrics returns 400, check that the counts aren't reset.
            m.count("foo", true)
            ck(400, 1, 0, <-metricsCalls)
    +       time.Sleep(100 * time.Millisecond)
            m.count("foo", false)
            ck(400, 1, 1, <-metricsCalls)
            m.count("foo", true)

The body sent to /client/metrics contains two timestamps. The de-duplication
done in the HTTP handler in the test therefore didn't work as expected. As a
fix, parse the request body in the HTTP handler and compare only the toggles
when doing the de-duplication.